+++ src/components/MainMapMaplibre.tsx
@@ -8,9 +8,7 @@ import InventorySource from "./map-components/InventorySource";
 import BaseLayerSource from "./map-components/BaseLayerSource";
 import UserLocationSource from "./map-components/UserLocationSource";
 import LayerInteraction from "./map-components/LayerInteraction";
-import { IonButton, IonIcon, IonLabel } from "@ionic/react";
-import { bookmarks, chevronDownOutline, options } from "ionicons/icons";
-import MapSelectionButton from "./MapSelectionButton";
+
 const MainMap: React.FC = () => {
   // onload callback handler
   const onLoad = (e: any) => {
