+++ src/components/map-components/InventorySource.tsx
@@ -1,20 +1,12 @@
 import { useEffect, useState } from "react";
 import cloneDeep from "lodash.clonedeep";
 import { CirclePaint, CircleLayout, MapLayerMouseEvent } from "mapbox-gl";
-import { Source, Layer, useMap, MapRef } from "react-map-gl";
+import { Source, Layer, useMap } from "react-map-gl";
 
 import { useData } from "../../context/data";
 import { InventoryData, InventoryFeature } from "../../context/data.model";
 import { useLayers } from "../../context/layers";
-import {
-  IonCard,
-  IonCardContent,
-  IonCardHeader,
-  IonCardTitle,
-  IonItem,
-  IonLabel,
-  IonPopover,
-} from "@ionic/react";
+import { IonPopover } from "@ionic/react";
 import { useHistory } from "react-router";
 import { useOffline } from "../../context/offline";
