+++ src/pages/TreeOverviewPage.tsx
@@ -9,8 +9,6 @@ import {
   IonToolbar,
 } from "@ionic/react";
 import { RouteComponentProps } from "react-router";
-
-
 import TreeDetails from "../components/TreeDetails";
 
 const TreeOverviewPage: React.FC<RouteComponentProps<{ id: string }>> = ({
